requirements/linux.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
Django==4.2.8
django-auditlog==2.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-import-export==3.3.5
django-parler==2.3
django-rosetta==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
setuptools==69.0.3
whitenoise==6.6.0